+++ src/anime_parsers_ru/parser_shikimori.py
@@ -22,11 +22,12 @@ class ShikimoriParser:
 
     genres_list = ['1-Action', '2-Adventure', '3-Racing', '4-Comedy', '5-Avant-Garde', '6-Mythology', '7-Mystery', '8-Drama', '9-Ecchi', '10-Fantasy', '11-Strategy-Game', '13-Historical', '14-Horror', '15-Kids', '17-Martial-Arts', '18-Mecha', '19-Music', '20-Parody', '21-Samurai', '22-Romance', '23-School', '24-Sci-Fi', '25-Shoujo', '27-Shounen', '29-Space', '30-Sports', '31-Super-Power', '32-Vampire', '35-Harem', '36-Slice-of-Life', '37-Supernatural', '38-Military', '39-Detective', '40-Psychological', '42-Seinen', '43-Josei', '102-Team-Sports', '103-Video-Game', '104-Adult-Cast', '105-Gore', '106-Reincarnation', '107-Love-Polygon', '108-Visual-Arts', '111-Time-Travel', '112-Gag-Humor', '114-Award-Winning', '117-Suspense', '118-Combat-Sports', '119-CGDCT', '124-Mahou-Shoujo', '125-Reverse-Harem', '130-Isekai', '131-Delinquents', '134-Childcare', '135-Magical-Sex-Shift', '136-Showbiz', '137-Otaku-Culture', '138-Organized-Crime', '139-Workplace', '140-Iyashikei', '141-Survival', '142-Performing-Arts', '143-Anthropomorphic', '144-Crossdressing', '145-Idols-(Female)', '146-High-Stakes-Game', '147-Medical', '148-Pets', '149-Educational', '150-Idols-(Male)', '151-Romantic-Subtext', '543-Gourmet']
 
-    def __init__(self, use_lxml: bool = False, mirror: str|None = None, proxy: str | None = None) -> None:
+    def __init__(self, use_lxml: bool = False, mirror: str|None = None, proxy: str | None = None, graphql_mirror: str | None = 'shikimori.io') -> None:
         """
         :use_lxml: Использовать lxml парсер. В некоторых случаях может не работать, однако работает значительно быстрее стандартного.
         :mirror: В случае, если оригинальный домен заблокирован, можно использовать этот параметр, чтобы заменить адрес сайта на зеркало. Пример: "1234.net"
         :proxy: http или socks5 прокси (`http://host:port` или `socks5://user:pass@host:port`) по умолчанию None
+        :graphql_mirror: Зеркало для graphql (по умолчанию shikimori.io)
         """
         if not LXML_WORKS and use_lxml:
             raise ImportWarning('Параметр use_lxml установлен в true, однако при попытке импорта lxml произошла ошибка')
@@ -36,6 +37,7 @@ class ShikimoriParser:
             self._dmn = mirror
         else:
             self._dmn = "shikimori.one"
+        self._graphql_mirror = graphql_mirror
 
     def search(self, title: str) -> list:
         """
@@ -696,12 +698,17 @@ class ShikimoriParser:
         return_query += '}'
         query += return_query + '\n}'
 
+        if self._graphql_mirror:
+            dmn = self._graphql_mirror
+        else:
+            dmn = self._dmn
+
         headers = {
             'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:128.0) Gecko/20100101 Firefox/128.0',
             'Accept': '*/*',
-            'Referer': f'https://{self._dmn}/api/doc/graphql',
+            'Referer': f'https://{dmn}/api/doc/graphql',
             'content-type': 'application/json',
-            'Origin': f'https://{self._dmn}',
+            'Origin': f'https://{dmn}',
         }
 
         json_data = {
@@ -710,7 +717,7 @@ class ShikimoriParser:
             'query': query
         }
         
-        response = requests.post(f'https://{self._dmn}/api/graphql', headers=headers, json=json_data, proxies=self.proxies)
+        response = requests.post(f'https://{dmn}/api/graphql', headers=headers, json=json_data, proxies=self.proxies)
         if response.status_code == 429:
             raise errors.TooManyRequests(f'Сервер вернул код 429 для обозначения что запросы выполняются слишком часто.')
         elif response.status_code == 520:
@@ -749,12 +756,17 @@ class ShikimoriParser:
         return_query += '}'
         query += return_query + '\n}'
 
+        if self._graphql_mirror:
+            dmn = self._graphql_mirror
+        else:
+            dmn = self._dmn
+
         headers = {
             'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:128.0) Gecko/20100101 Firefox/128.0',
             'Accept': '*/*',
-            'Referer': f'https://{self._dmn}/api/doc/graphql',
+            'Referer': f'https://{dmn}/api/doc/graphql',
             'content-type': 'application/json',
-            'Origin': f'https://{self._dmn}',
+            'Origin': f'https://{dmn}',
         }
 
         json_data = {
@@ -763,7 +775,7 @@ class ShikimoriParser:
             'query': query
         }
         
-        response = requests.post(f'https://{self._dmn}/api/graphql', headers=headers, json=json_data, proxies=self.proxies)
+        response = requests.post(f'https://{dmn}/api/graphql', headers=headers, json=json_data, proxies=self.proxies)
         if response.status_code == 429:
             raise errors.TooManyRequests(f'Сервер вернул код 429 для обозначения что запросы выполняются слишком часто.')
         elif response.status_code == 520: